pyproject.toml
[project]
name = "ragpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.0",
    "openai>=2.14.0",
    "pyinstaller>=6.17.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
